fix socket function

+++ fin-crawling-fast-api/server/app/crawler/MarcapCrawler.py
@@ -17,9 +17,9 @@ import uuid
 from app.observer.DownloadObserver import DownloadObserver
 from app.observer.CmdFileSystemEventHandler import FILE_SYSTEM_HANDLER
 
-from .Crawler import Crawler
 from pathlib import Path
-from app.model.dto import StockCrawlingDownloadTaskDTO, StockCrawlingRunCrawlingDTO, StockCrawlingTaskDTO, StockMarketCapitalResultDTO, StockMarketCapitalDTO
+from app.model.dto import StockCrawlingDownloadTaskDTO, StockCrawlingRunCrawlingDTO, StockMarketCapitalResultDTO, StockMarketCapitalDTO
+from uvicorn.config import logger
 
 T = TypeVar("T")
 
@@ -30,12 +30,12 @@ EVENT_MARCAP_CRAWLING_ON_DOWNLOAD_COMPLETE: Final = "marcapCrawler/onDownloadCom
 EVENT_MARCAP_CRAWLING_ON_PARSING_COMPLETE: Final = "marcapCrawler/onParsingComplete"
 
 
-class MarcapCrawler(Crawler[T]):
-    @staticmethod
-    def create(ee: EventEmitter) -> MarcapCrawler:
-        crawler = MarcapCrawler[None]()
-        crawler.ee = EventEmitter()
-        return crawler
+class MarcapCrawler(object):
+    
+    def __init__(self) -> None:
+        super().__init__()
+        self.ee = EventEmitter()
+        self.isLock = False
 
     def createUUID(self) -> str:
         return str(uuid.uuid4())
@@ -44,7 +44,7 @@ class MarcapCrawler(Crawler[T]):
         chrome_options = webdriver.ChromeOptions()
         prefs = {
             'profile.default_content_setting_values.automatic_downloads': 1,
-            'download.default_directory': f"/usr/src/downloads/{uuid}"
+            'download.default_directory': f"/home/seluser/Downloads/{uuid}"
         }
         chrome_options.add_experimental_option("prefs", prefs)
         driver = webdriver.Remote(
@@ -56,6 +56,7 @@ class MarcapCrawler(Crawler[T]):
     async def crawling(self, dto: StockCrawlingRunCrawlingDTO) -> None:
         driver = None
         uuid = self.createUUID()
+        logger.info(uuid)
         self.ee.emit(EVENT_MARCAP_CRAWLING_ON_CONNECTING_WEBDRIVER, dto)
         try:
             driver = self.connectWebDriver(dto.driverAddr, uuid)
@@ -71,6 +72,8 @@ class MarcapCrawler(Crawler[T]):
             WebDriverWait(driver, timeout=20, poll_frequency=1).until(EC.element_to_be_clickable((By.CSS_SELECTOR, "#mktId_0_1")))
             date = datetime.strptime(dto.startDateStr, "%Y%m%d")
             endDate = datetime.strptime(dto.endDateStr, "%Y%m%d")
+            downloadObserver = DownloadObserver()
+            downloadObserver.startObserver(uuid, self.ee)
 
             while date <= endDate:
                 dateStr = date.strftime("%Y%m%d")
@@ -81,18 +84,19 @@ class MarcapCrawler(Crawler[T]):
                     "taskId": dto.taskId,
                     "taskUniqueId": dto.taskUniqueId
                 })
+                downloadObserver.event_handler.setDownloadTask(downloadTask)
                 self.ee.emit(EVENT_MARCAP_CRAWLING_ON_DOWNLOAD_START, downloadTask)
-                downloadObserver = DownloadObserver()
-                downloadObserver.startObserver(uuid, self.ee)
                 await self.downloadData(downloadTask, downloadObserver, driver)
                 date = date + timedelta(days=1)
         except Exception as e:
-            print(f"error: {str(e)}")
+            logger.error(f"error: {str(e)}")
         finally:
+            downloadObserver.stopObserver()
             if driver:
                 driver.quit()
     
     async def downloadData(self, downloadTask: StockCrawlingDownloadTaskDTO, downloadObserver: DownloadObserver, driver: WebDriver) -> None:
+        logger.info("downloadData")
         if driver is None:
             return
         # pymitter
@@ -113,25 +117,27 @@ class MarcapCrawler(Crawler[T]):
             after = driver.execute_script('return $(".CI-MDI-UNIT-TIME").text()')
             await asyncio.sleep(0.5)
         #     driver.implicitly_wait(1)
-        print("before:"+before)
-        print("after:"+after)
+        logger.info("before:"+before)
+        logger.info("after:"+after)
         await asyncio.sleep(3)
         WebDriverWait(driver, timeout=10, poll_frequency=1).until(EC.element_to_be_clickable((By.CSS_SELECTOR, "*[title='다운로드 팝업']")))
         driver.execute_script("$('[title=\"다운로드 팝업\"]').click()")
         WebDriverWait(driver, timeout=10, poll_frequency=1).until(EC.element_to_be_clickable((By.CSS_SELECTOR, "*[data-type='csv']")))
         driver.execute_script("$(\"[data-type='csv']\").click()")
         print("wait:"+downloadTask.dateStr)
+        logger.info("wait:"+downloadTask.dateStr)
         self.isLock = True
 
-        @self.ee.on(FILE_SYSTEM_HANDLER(downloadTask.uuid))
+        mainThreadLoop = asyncio.get_running_loop()
+
+        @self.ee.once(FILE_SYSTEM_HANDLER(downloadTask.uuid))
         def downloadComplete(event: FileCreatedEvent, downloadTask: StockCrawlingDownloadTaskDTO) -> None:
-            self.isLock = False
             self.ee.emit(EVENT_MARCAP_CRAWLING_ON_DOWNLOAD_COMPLETE, downloadTask)
-            self.parseFile(event, downloadTask, downloadObserver)
+            mainThreadLoop.create_task(self.parseFile(event, downloadTask, downloadObserver))
 
         while self.isLock:
             await asyncio.sleep(1)
-            print(f"isLock:{str(self.isLock)}")
+            logger.info(f"isLock:{str(self.isLock)}")
     
     def convertFileToDto(self, path: str, dto: StockMarketCapitalResultDTO) -> None:
         lines = []
@@ -141,24 +147,24 @@ class MarcapCrawler(Crawler[T]):
             lines = f.readlines()
         
         for i in range(1, len(lines)):
-            marcap = StockMarketCapitalDTO()
             data = lines[i].replace('"', '').split(",")
-            marcap.date = dto.date
-            marcap.market = dto.market
-            marcap.code = data[0].strip()
-            marcap.name = data[1].strip()
-            marcap.close = data[2].strip()
-            marcap.diff = data[3].strip()
-            marcap.percent = data[4].strip()
-            marcap.open = data[5].strip()
-            marcap.high = data[6].strip()
-            marcap.low = data[7].strip()
-            marcap.volume = data[8].strip()
-            marcap.price = data[9].strip()
-            marcap.marcap = data[10].strip()
-            marcap.number = data[11].strip()
-            marcap.createdAt = datetime.today()
-            marcap.updatedAt = ""
+            marcap = StockMarketCapitalDTO(**{
+                "date": dto.date,
+                "market": dto.market,
+                "code": data[0].strip(),
+                "name": data[1].strip(),
+                "close": data[2].strip(),
+                "diff": data[3].strip(),
+                "percent": data[4].strip(),
+                "open": data[5].strip(),
+                "high": data[6].strip(),
+                "low": data[7].strip(),
+                "volume": data[8].strip(),
+                "price": data[9].strip(),
+                "marcap": data[10].strip(),
+                "number": data[11].strip()
+            })
+            
             dto.data.append(marcap)
 
     async def isExistFile(self, path: str, ext: str = ".csv") -> bool:
@@ -180,21 +186,27 @@ class MarcapCrawler(Crawler[T]):
             retdto.market = market
             isExist = await self.isExistFile(event.src_path)
             if not isExist:
-                raise Exception('file Not Exist')
+                return
             print("created: " + date)
             await asyncio.sleep(0.5)
             dest_path = f'{os.path.dirname(event.src_path)}/{date}.csv'
+            if os.path.isfile(dest_path):
+                return
             self.changeCharSet(event.src_path)
             os.rename(event.src_path, dest_path)
             self.convertFileToDto(dest_path, retdto)
             retdto.result = "success"
             isSuccess = True
+            self.ee.emit(EVENT_MARCAP_CRAWLING_ON_PARSING_COMPLETE, isSuccess, retdto, downloadTask)
+            self.isLock = False
         except Exception as e:
             retdto.result = "fail"
             retdto.errorMsg = str(e)
-        finally:
-            downloadObserver.stopObserver()
+            logger.error(str(e))
             self.ee.emit(EVENT_MARCAP_CRAWLING_ON_PARSING_COMPLETE, isSuccess, retdto, downloadTask)
+            self.isLock = False
+        finally:
+            logger.info("parseFile...")
 
     def changeCharSet(self, path: str) -> None:
         lines = None
@@ -203,6 +215,3 @@ class MarcapCrawler(Crawler[T]):
         with open(path, 'w', encoding="utf-8") as f:
             f.writelines(lines)
 
-
-ee = EventEmitter()
-marcapCrawler: MarcapCrawler = MarcapCrawler(ee)
